+++ src/components.d.ts
@@ -50,6 +50,11 @@ export namespace Components {
     interface IfxCardText {
         "hasBtn": boolean;
     }
+    interface IfxCheckbox {
+        "checked": boolean;
+        "disabled": boolean;
+        "error": boolean;
+    }
     interface IfxDropdown {
         "disabled": boolean;
         "filter": boolean;
@@ -82,6 +87,11 @@ export namespace Components {
         "search": boolean;
         "size": 's' | 'm';
     }
+    interface IfxFooter {
+        "variant": string;
+    }
+    interface IfxFooterColumn {
+    }
     interface IfxIcon {
         "icon": any;
         "ifxIcon": any;
@@ -97,10 +107,12 @@ export namespace Components {
         "target": string;
         "variant": 'solid' | 'outline' | 'outline-text';
     }
+    interface IfxIconsPreview {
+    }
     interface IfxLink {
+        "bold": boolean;
+        "color": string;
         "href": string;
-        "icon": any;
-        "position": string;
         "target": string;
         "underline": any;
     }
@@ -119,6 +131,14 @@ export namespace Components {
         "isFlush": boolean;
         "titleText": string;
     }
+    interface IfxNavbar {
+        "applicationName": string;
+    }
+    interface IfxNavbarMenuItem {
+        "hideLabel": boolean;
+        "icon": string;
+        "isProfile": boolean;
+    }
     interface IfxNumberIndicator {
         "inverted": boolean;
     }
@@ -127,6 +147,11 @@ export namespace Components {
         "percentage": number;
         "showLabel": boolean;
         "size": string;
+    }
+    interface IfxRadioButton {
+        "checked": boolean;
+        "disabled": boolean;
+        "error": boolean;
     }
     interface IfxRange {
         "disabled": boolean;
@@ -140,10 +165,11 @@ export namespace Components {
         "value": number;
     }
     interface IfxSearchBar {
+        "hideLabel": boolean;
         "icon": string;
         "isOpen": boolean;
         "showCloseButton": boolean;
-        "width": string;
+        "size": string;
     }
     interface IfxSearchInput {
         "borderColor": 'light' | 'dark' | 'green';
@@ -171,6 +197,56 @@ export namespace Components {
     interface IfxTag {
         "text": string;
     }
+    interface IfxToggle {
+        "checked": boolean;
+    }
+    interface InfineonIconStencil {
+        "icon": any;
+    }
+}
+export interface IfxAccordionItemCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLIfxAccordionItemElement;
+}
+export interface IfxDropdownItemCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLIfxDropdownItemElement;
+}
+export interface IfxDropdownMenuCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLIfxDropdownMenuElement;
+}
+export interface IfxFilterInputCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLIfxFilterInputElement;
+}
+export interface IfxProgressBarCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLIfxProgressBarElement;
+}
+export interface IfxRangeCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLIfxRangeElement;
+}
+export interface IfxSearchBarCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLIfxSearchBarElement;
+}
+export interface IfxSearchInputCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLIfxSearchInputElement;
+}
+export interface IfxTabCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLIfxTabElement;
+}
+export interface IfxTabsCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLIfxTabsElement;
+}
+export interface IfxToggleCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLIfxToggleElement;
 }
 declare global {
     interface HTMLIfxAccordionElement extends Components.IfxAccordion, HTMLStencilElement {
@@ -227,6 +303,12 @@ declare global {
         prototype: HTMLIfxCardTextElement;
         new (): HTMLIfxCardTextElement;
     };
+    interface HTMLIfxCheckboxElement extends Components.IfxCheckbox, HTMLStencilElement {
+    }
+    var HTMLIfxCheckboxElement: {
+        prototype: HTMLIfxCheckboxElement;
+        new (): HTMLIfxCheckboxElement;
+    };
     interface HTMLIfxDropdownElement extends Components.IfxDropdown, HTMLStencilElement {
     }
     var HTMLIfxDropdownElement: {
@@ -251,6 +333,18 @@ declare global {
         prototype: HTMLIfxFilterInputElement;
         new (): HTMLIfxFilterInputElement;
     };
+    interface HTMLIfxFooterElement extends Components.IfxFooter, HTMLStencilElement {
+    }
+    var HTMLIfxFooterElement: {
+        prototype: HTMLIfxFooterElement;
+        new (): HTMLIfxFooterElement;
+    };
+    interface HTMLIfxFooterColumnElement extends Components.IfxFooterColumn, HTMLStencilElement {
+    }
+    var HTMLIfxFooterColumnElement: {
+        prototype: HTMLIfxFooterColumnElement;
+        new (): HTMLIfxFooterColumnElement;
+    };
     interface HTMLIfxIconElement extends Components.IfxIcon, HTMLStencilElement {
     }
     var HTMLIfxIconElement: {
@@ -262,6 +356,12 @@ declare global {
     var HTMLIfxIconButtonElement: {
         prototype: HTMLIfxIconButtonElement;
         new (): HTMLIfxIconButtonElement;
+    };
+    interface HTMLIfxIconsPreviewElement extends Components.IfxIconsPreview, HTMLStencilElement {
+    }
+    var HTMLIfxIconsPreviewElement: {
+        prototype: HTMLIfxIconsPreviewElement;
+        new (): HTMLIfxIconsPreviewElement;
     };
     interface HTMLIfxLinkElement extends Components.IfxLink, HTMLStencilElement {
     }
@@ -287,6 +387,18 @@ declare global {
         prototype: HTMLIfxListNotificationElement;
         new (): HTMLIfxListNotificationElement;
     };
+    interface HTMLIfxNavbarElement extends Components.IfxNavbar, HTMLStencilElement {
+    }
+    var HTMLIfxNavbarElement: {
+        prototype: HTMLIfxNavbarElement;
+        new (): HTMLIfxNavbarElement;
+    };
+    interface HTMLIfxNavbarMenuItemElement extends Components.IfxNavbarMenuItem, HTMLStencilElement {
+    }
+    var HTMLIfxNavbarMenuItemElement: {
+        prototype: HTMLIfxNavbarMenuItemElement;
+        new (): HTMLIfxNavbarMenuItemElement;
+    };
     interface HTMLIfxNumberIndicatorElement extends Components.IfxNumberIndicator, HTMLStencilElement {
     }
     var HTMLIfxNumberIndicatorElement: {
@@ -298,6 +410,12 @@ declare global {
     var HTMLIfxProgressBarElement: {
         prototype: HTMLIfxProgressBarElement;
         new (): HTMLIfxProgressBarElement;
+    };
+    interface HTMLIfxRadioButtonElement extends Components.IfxRadioButton, HTMLStencilElement {
+    }
+    var HTMLIfxRadioButtonElement: {
+        prototype: HTMLIfxRadioButtonElement;
+        new (): HTMLIfxRadioButtonElement;
     };
     interface HTMLIfxRangeElement extends Components.IfxRange, HTMLStencilElement {
     }
@@ -347,7 +465,17 @@ declare global {
         prototype: HTMLIfxTagElement;
         new (): HTMLIfxTagElement;
     };
-
+    interface HTMLIfxToggleElement extends Components.IfxToggle, HTMLStencilElement {
+    }
+    var HTMLIfxToggleElement: {
+        prototype: HTMLIfxToggleElement;
+        new (): HTMLIfxToggleElement;
+    };
+    interface HTMLInfineonIconStencilElement extends Components.InfineonIconStencil, HTMLStencilElement {
+    }
+    var HTMLInfineonIconStencilElement: {
+        prototype: HTMLInfineonIconStencilElement;
+        new (): HTMLInfineonIconStencilElement;
     };
     interface HTMLElementTagNameMap {
         "ifx-accordion": HTMLIfxAccordionElement;
@@ -359,18 +487,25 @@ declare global {
         "ifx-card-headline": HTMLIfxCardHeadlineElement;
         "ifx-card-overline": HTMLIfxCardOverlineElement;
         "ifx-card-text": HTMLIfxCardTextElement;
+        "ifx-checkbox": HTMLIfxCheckboxElement;
         "ifx-dropdown": HTMLIfxDropdownElement;
         "ifx-dropdown-item": HTMLIfxDropdownItemElement;
         "ifx-dropdown-menu": HTMLIfxDropdownMenuElement;
         "ifx-filter-input": HTMLIfxFilterInputElement;
+        "ifx-footer": HTMLIfxFooterElement;
+        "ifx-footer-column": HTMLIfxFooterColumnElement;
         "ifx-icon": HTMLIfxIconElement;
         "ifx-icon-button": HTMLIfxIconButtonElement;
+        "ifx-icons-preview": HTMLIfxIconsPreviewElement;
         "ifx-link": HTMLIfxLinkElement;
         "ifx-list-group": HTMLIfxListGroupElement;
         "ifx-list-item": HTMLIfxListItemElement;
         "ifx-list-notification": HTMLIfxListNotificationElement;
+        "ifx-navbar": HTMLIfxNavbarElement;
+        "ifx-navbar-menu-item": HTMLIfxNavbarMenuItemElement;
         "ifx-number-indicator": HTMLIfxNumberIndicatorElement;
         "ifx-progress-bar": HTMLIfxProgressBarElement;
+        "ifx-radio-button": HTMLIfxRadioButtonElement;
         "ifx-range": HTMLIfxRangeElement;
         "ifx-search-bar": HTMLIfxSearchBarElement;
         "ifx-search-input": HTMLIfxSearchInputElement;
@@ -379,7 +514,8 @@ declare global {
         "ifx-tab": HTMLIfxTabElement;
         "ifx-tabs": HTMLIfxTabsElement;
         "ifx-tag": HTMLIfxTagElement;
-
+        "ifx-toggle": HTMLIfxToggleElement;
+        "infineon-icon-stencil": HTMLInfineonIconStencilElement;
     }
 }
 declare namespace LocalJSX {
@@ -425,6 +561,11 @@ declare namespace LocalJSX {
     interface IfxCardText {
         "hasBtn"?: boolean;
     }
+    interface IfxCheckbox {
+        "checked"?: boolean;
+        "disabled"?: boolean;
+        "error"?: boolean;
+    }
     interface IfxDropdown {
         "disabled"?: boolean;
         "filter"?: boolean;
@@ -460,6 +601,11 @@ declare namespace LocalJSX {
         "search"?: boolean;
         "size"?: 's' | 'm';
     }
+    interface IfxFooter {
+        "variant"?: string;
+    }
+    interface IfxFooterColumn {
+    }
     interface IfxIcon {
         "icon"?: any;
         "ifxIcon"?: any;
@@ -474,10 +620,12 @@ declare namespace LocalJSX {
         "target"?: string;
         "variant"?: 'solid' | 'outline' | 'outline-text';
     }
+    interface IfxIconsPreview {
+    }
     interface IfxLink {
+        "bold"?: boolean;
+        "color"?: string;
         "href"?: string;
-        "icon"?: any;
-        "position"?: string;
         "target"?: string;
         "underline"?: any;
     }
@@ -496,15 +644,28 @@ declare namespace LocalJSX {
         "isFlush"?: boolean;
         "titleText"?: string;
     }
+    interface IfxNavbar {
+        "applicationName"?: string;
+    }
+    interface IfxNavbarMenuItem {
+        "hideLabel"?: boolean;
+        "icon"?: string;
+        "isProfile"?: boolean;
+    }
     interface IfxNumberIndicator {
         "inverted"?: boolean;
     }
     interface IfxProgressBar {
         "label"?: string;
-        "onPercentageChange"?: (event: IfxProgressBarCustomEvent<number>) => void;
+        "onPercentageChange"?: (event: IfxProgressBarCustomEvent<CustomEvent>) => void;
         "percentage"?: number;
         "showLabel"?: boolean;
         "size"?: string;
+    }
+    interface IfxRadioButton {
+        "checked"?: boolean;
+        "disabled"?: boolean;
+        "error"?: boolean;
     }
     interface IfxRange {
         "disabled"?: boolean;
@@ -519,11 +680,12 @@ declare namespace LocalJSX {
         "value"?: number;
     }
     interface IfxSearchBar {
+        "hideLabel"?: boolean;
         "icon"?: string;
         "isOpen"?: boolean;
         "onSearch"?: (event: IfxSearchBarCustomEvent<string>) => void;
         "showCloseButton"?: boolean;
-        "width"?: string;
+        "size"?: string;
     }
     interface IfxSearchInput {
         "borderColor"?: 'light' | 'dark' | 'green';
@@ -555,7 +717,12 @@ declare namespace LocalJSX {
     interface IfxTag {
         "text"?: string;
     }
-
+    interface IfxToggle {
+        "checked"?: boolean;
+        "onValueChanged"?: (event: IfxToggleCustomEvent<boolean>) => void;
+    }
+    interface InfineonIconStencil {
+        "icon"?: any;
     }
     interface IntrinsicElements {
         "ifx-accordion": IfxAccordion;
@@ -567,18 +734,25 @@ declare namespace LocalJSX {
         "ifx-card-headline": IfxCardHeadline;
         "ifx-card-overline": IfxCardOverline;
         "ifx-card-text": IfxCardText;
+        "ifx-checkbox": IfxCheckbox;
         "ifx-dropdown": IfxDropdown;
         "ifx-dropdown-item": IfxDropdownItem;
         "ifx-dropdown-menu": IfxDropdownMenu;
         "ifx-filter-input": IfxFilterInput;
+        "ifx-footer": IfxFooter;
+        "ifx-footer-column": IfxFooterColumn;
         "ifx-icon": IfxIcon;
         "ifx-icon-button": IfxIconButton;
+        "ifx-icons-preview": IfxIconsPreview;
         "ifx-link": IfxLink;
         "ifx-list-group": IfxListGroup;
         "ifx-list-item": IfxListItem;
         "ifx-list-notification": IfxListNotification;
+        "ifx-navbar": IfxNavbar;
+        "ifx-navbar-menu-item": IfxNavbarMenuItem;
         "ifx-number-indicator": IfxNumberIndicator;
         "ifx-progress-bar": IfxProgressBar;
+        "ifx-radio-button": IfxRadioButton;
         "ifx-range": IfxRange;
         "ifx-search-bar": IfxSearchBar;
         "ifx-search-input": IfxSearchInput;
@@ -587,7 +761,8 @@ declare namespace LocalJSX {
         "ifx-tab": IfxTab;
         "ifx-tabs": IfxTabs;
         "ifx-tag": IfxTag;
-
+        "ifx-toggle": IfxToggle;
+        "infineon-icon-stencil": InfineonIconStencil;
     }
 }
 export { LocalJSX as JSX };
@@ -603,18 +778,25 @@ declare module "@stencil/core" {
             "ifx-card-headline": LocalJSX.IfxCardHeadline & JSXBase.HTMLAttributes<HTMLIfxCardHeadlineElement>;
             "ifx-card-overline": LocalJSX.IfxCardOverline & JSXBase.HTMLAttributes<HTMLIfxCardOverlineElement>;
             "ifx-card-text": LocalJSX.IfxCardText & JSXBase.HTMLAttributes<HTMLIfxCardTextElement>;
+            "ifx-checkbox": LocalJSX.IfxCheckbox & JSXBase.HTMLAttributes<HTMLIfxCheckboxElement>;
             "ifx-dropdown": LocalJSX.IfxDropdown & JSXBase.HTMLAttributes<HTMLIfxDropdownElement>;
             "ifx-dropdown-item": LocalJSX.IfxDropdownItem & JSXBase.HTMLAttributes<HTMLIfxDropdownItemElement>;
             "ifx-dropdown-menu": LocalJSX.IfxDropdownMenu & JSXBase.HTMLAttributes<HTMLIfxDropdownMenuElement>;
             "ifx-filter-input": LocalJSX.IfxFilterInput & JSXBase.HTMLAttributes<HTMLIfxFilterInputElement>;
+            "ifx-footer": LocalJSX.IfxFooter & JSXBase.HTMLAttributes<HTMLIfxFooterElement>;
+            "ifx-footer-column": LocalJSX.IfxFooterColumn & JSXBase.HTMLAttributes<HTMLIfxFooterColumnElement>;
             "ifx-icon": LocalJSX.IfxIcon & JSXBase.HTMLAttributes<HTMLIfxIconElement>;
             "ifx-icon-button": LocalJSX.IfxIconButton & JSXBase.HTMLAttributes<HTMLIfxIconButtonElement>;
+            "ifx-icons-preview": LocalJSX.IfxIconsPreview & JSXBase.HTMLAttributes<HTMLIfxIconsPreviewElement>;
             "ifx-link": LocalJSX.IfxLink & JSXBase.HTMLAttributes<HTMLIfxLinkElement>;
             "ifx-list-group": LocalJSX.IfxListGroup & JSXBase.HTMLAttributes<HTMLIfxListGroupElement>;
             "ifx-list-item": LocalJSX.IfxListItem & JSXBase.HTMLAttributes<HTMLIfxListItemElement>;
             "ifx-list-notification": LocalJSX.IfxListNotification & JSXBase.HTMLAttributes<HTMLIfxListNotificationElement>;
+            "ifx-navbar": LocalJSX.IfxNavbar & JSXBase.HTMLAttributes<HTMLIfxNavbarElement>;
+            "ifx-navbar-menu-item": LocalJSX.IfxNavbarMenuItem & JSXBase.HTMLAttributes<HTMLIfxNavbarMenuItemElement>;
             "ifx-number-indicator": LocalJSX.IfxNumberIndicator & JSXBase.HTMLAttributes<HTMLIfxNumberIndicatorElement>;
             "ifx-progress-bar": LocalJSX.IfxProgressBar & JSXBase.HTMLAttributes<HTMLIfxProgressBarElement>;
+            "ifx-radio-button": LocalJSX.IfxRadioButton & JSXBase.HTMLAttributes<HTMLIfxRadioButtonElement>;
             "ifx-range": LocalJSX.IfxRange & JSXBase.HTMLAttributes<HTMLIfxRangeElement>;
             "ifx-search-bar": LocalJSX.IfxSearchBar & JSXBase.HTMLAttributes<HTMLIfxSearchBarElement>;
             "ifx-search-input": LocalJSX.IfxSearchInput & JSXBase.HTMLAttributes<HTMLIfxSearchInputElement>;
@@ -623,7 +805,8 @@ declare module "@stencil/core" {
             "ifx-tab": LocalJSX.IfxTab & JSXBase.HTMLAttributes<HTMLIfxTabElement>;
             "ifx-tabs": LocalJSX.IfxTabs & JSXBase.HTMLAttributes<HTMLIfxTabsElement>;
             "ifx-tag": LocalJSX.IfxTag & JSXBase.HTMLAttributes<HTMLIfxTagElement>;
-
+            "ifx-toggle": LocalJSX.IfxToggle & JSXBase.HTMLAttributes<HTMLIfxToggleElement>;
+            "infineon-icon-stencil": LocalJSX.InfineonIconStencil & JSXBase.HTMLAttributes<HTMLInfineonIconStencilElement>;
         }
     }
 }