+++ src/components/tabs/tabs.stories.tsx
@@ -1,6 +1,5 @@
 import { action } from '@storybook/addon-actions';
 
-
 export default {
   title: 'Components/Tabs',
   component: 'ifx-tabs',